+++ Carbon/Functions/Test-CCryptoKeyAvailable.ps1
@@ -1,4 +1,0 @@
-function Test-CCryptoKeyAvailable
-{
-    return $null -ne [Type]::GetType('System.Security.AccessControl.CryptoKeyRights')
-}