+++ Start-CarbonTest.ps1
@@ -10,9 +10,9 @@ Run `Complete-CarbonTest.ps1` to reverse the changes this script makes.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #     http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -33,9 +33,9 @@ Install-CJunction -Link $carbonModuleRoot -Target (Join-Path -Path $PSScriptRoot
 
 if( (Test-Path -Path 'env:APPVEYOR') )
 {
-    Grant-Permission -Path ($PSScriptRoot | Split-Path) -Identity 'Everyone' -Permission 'FullControl'
-    Grant-Permission -Path ('C:\Users\appveyor\Documents') -Identity 'Everyone' -Permission 'FullControl'
-
+    Grant-Permission -Path ($PSScriptRoot | Split-Path) -Identity 'Everyone' -Permission 'FullControl' -NoWarn
+    Grant-Permission -Path ('C:\Users\appveyor\Documents') -Identity 'Everyone' -Permission 'FullControl' -NoWarn
+ -NoWarn
     $wmiprvse = Get-Process -Name 'wmiprvse'
     #$wmiprvse | Format-Table
     $wmiprvse | Stop-Process -Force
@@ -54,7 +54,7 @@ configuration Yolo
             }
 
             SetScript = {
-            
+
             }
 
             TestScript =  {