+++ Test/Security/Test-ConvertToContainerInheritanceFlags.ps1
@@ -1,9 +1,9 @@
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
-# 
+#
 #     http://www.apache.org/licenses/LICENSE-2.0
-# 
+#
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -28,9 +28,9 @@ function Test-ShouldConvertToNtfsContainerInheritanceFlags
     try
     {
         [Enum]::GetValues([Carbon.Security.ContainerInheritanceFlags]) | ForEach-Object {
-            Grant-Permission -Path $tempDir -Identity $UserName -Permission FullControl -ApplyTo $_
-            $perm = Get-Permission -Path $tempDir -Identity $UserName
-            $flags = ConvertTo-ContainerInheritanceFlags -InheritanceFlags $perm.InheritanceFlags -PropagationFlags $perm.PropagationFlags
+            Grant-Permission -Path $tempDir -Identity $UserName -Permission FullControl -ApplyTo $_ -NoWarn
+            $perm = Get-Permission -Path $tempDir -Identity $UserName -NoWarn
+            $flags = ConvertTo-ContainerInheritanceFlags -InheritanceFlags $perm.InheritanceFlags -PropagationFlags $perm.PropagationFlags -NoWarn
             Assert-Equal $_ $flags
         }
     }
@@ -47,14 +47,14 @@ function Test-ShouldConvertToRegistryContainerInheritanceFlags
 {
     $tempDir = 'Carbon+{0}+{1}' -f ((Split-Path -Leaf -Path $PSCommandPath),([IO.Path]::GetRandomFileName()))
     $tempDir = Join-Path -Path 'hkcu:\' -ChildPath $tempDir
-    New-Item -Path $tempDir 
+    New-Item -Path $tempDir
 
     try
     {
         [Enum]::GetValues([Carbon.Security.ContainerInheritanceFlags]) | ForEach-Object {
-            Grant-Permission -Path $tempDir -Identity $UserName -Permission ReadKey -ApplyTo $_
-            $perm = Get-Permission -Path $tempDir -Identity $UserName
-            $flags = ConvertTo-ContainerInheritanceFlags -InheritanceFlags $perm.InheritanceFlags -PropagationFlags $perm.PropagationFlags
+            Grant-Permission -Path $tempDir -Identity $UserName -Permission ReadKey -ApplyTo $_ -NoWarn
+            $perm = Get-Permission -Path $tempDir -Identity $UserName -NoWarn
+            $flags = ConvertTo-ContainerInheritanceFlags -InheritanceFlags $perm.InheritanceFlags -PropagationFlags $perm.PropagationFlags -NoWarn
             Assert-Equal $_ $flags
         }
     }