+++ Test/Test-CPermission.Tests.ps1
@@ -28,7 +28,11 @@ BeforeAll {
     $script:dirPath = Join-Path -Path $script:tempDir -ChildPath 'Directory'
     $script:filePath = Join-Path -Path $script:dirPath -ChildPath 'File'
     New-Item -Path $script:filePath -ItemType File
-    Grant-CPermission -Identity $script:identity -Permission ReadAndExecute -Path $script:dirPath -ApplyTo 'ChildLeaves'
+    Grant-CPermission -Identity $script:identity `
+                      -Permission ReadAndExecute `
+                      -Path $script:dirPath `
+                      -NoWarn `
+                      -ApplyTo 'ChildLeaves'
 
     $script:tempKeyPath = 'hkcu:\Software\Carbon\Test'
     $script:keyPath = Join-Path -Path $script:tempKeyPath -ChildPath 'Test-CPermission'
@@ -37,7 +41,8 @@ BeforeAll {
     Grant-CPermission -Identity $script:identity `
                       -Permission 'ReadKey','WriteKey' `
                       -Path $script:keyPath `
-                      -ApplyTo 'ChildLeaves'
+                      -ApplyTo 'ChildLeaves' `
+                      -NoWarn
 
     $script:testDirPermArgs = @{
         Path = $script:dirPath;
@@ -62,41 +67,41 @@ Describe 'Test-CPermission' {
     }
 
     It 'should handle non existent path' {
-        Test-CPermission -Path 'C:\I\Do\Not\Exist' -Identity $script:identity -Permission 'FullControl' -ErrorAction SilentlyContinue |
+        Test-CPermission -Path 'C:\I\Do\Not\Exist' -Identity $script:identity -Permission 'FullControl' -NoWarn -ErrorAction SilentlyContinue |
             Should -BeNullOrEmpty
         $Global:Error | Should -HaveCount 2
     }
 
     It 'should check ungranted permission on file system' {
-        Test-CPermission @testDirPermArgs -Permission 'Write' | Should -BeFalse
+        Test-CPermission @testDirPermArgs -Permission 'Write' -NoWarn | Should -BeFalse
     }
 
     It 'should check granted permission on file system' {
-        Test-CPermission @testDirPermArgs -Permission 'Read' | Should -BeTrue
+        Test-CPermission @testDirPermArgs -Permission 'Read' -NoWarn | Should -BeTrue
     }
 
     It 'should check exact partial permission on file system' {
-        Test-CPermission @testDirPermArgs -Permission 'Read' -Exact | Should -BeFalse
+        Test-CPermission @testDirPermArgs -Permission 'Read' -Exact -NoWarn | Should -BeFalse
     }
 
     It 'should check exact permission on file system' {
-        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -Exact | Should -BeTrue
+        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -Exact -NoWarn | Should -BeTrue
     }
 
     It 'should exclude inherited permission' {
-        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' | Should -BeFalse
+        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -NoWarn | Should -BeFalse
     }
 
     It 'should include inherited permission' {
-        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -Inherited | Should -BeTrue
+        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -Inherited -NoWarn | Should -BeTrue
     }
 
     It 'should exclude inherited partial permission' {
-        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -Exact | Should -BeFalse
+        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -Exact -NoWarn | Should -BeFalse
     }
 
     It 'should include inherited exact permission' {
-        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -Inherited -Exact | Should -BeTrue
+        Test-CPermission @testFilePermArgs -Permission 'ReadAndExecute' -Inherited -Exact -NoWarn | Should -BeTrue
     }
 
     It 'should ignore inheritance and propagation flags on file' {
@@ -106,47 +111,49 @@ Describe 'Test-CPermission' {
                          -ApplyTo SubContainers `
                          -Inherited `
                          -WarningVariable 'warning' `
-                         -WarningAction SilentlyContinue |
+                         -WarningAction SilentlyContinue `
+                         -NoWarn |
             Should -BeTrue
         $warning | Should -Not -BeNullOrEmpty
         $warning[0] | Should -BeLike 'Can''t test inheritance/propagation rules on a leaf.*'
     }
 
     It 'should check ungranted permission on registry' {
-        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'Delete' | Should -BeFalse
+        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'Delete' -NoWarn | Should -BeFalse
     }
 
     It 'should check granted permission on registry' {
-        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'ReadKey' | Should -BeTrue
+        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'ReadKey' -NoWarn | Should -BeTrue
     }
 
     It 'should check exact partial permission on registry' {
-        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'ReadKey' -Exact | Should -BeFalse
+        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'ReadKey' -Exact -NoWarn | Should -BeFalse
     }
 
     It 'should check exact permission on registry' {
-        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'ReadKey','WriteKey' -Exact |
+        Test-CPermission -Path $script:keyPath -Identity $script:identity -Permission 'ReadKey','WriteKey' -Exact -NoWarn |
             Should -BeTrue
     }
 
     It 'should check ungranted inheritance flags' {
-        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ContainerAndSubContainersAndLeaves  |
+        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ContainerAndSubContainersAndLeaves -NoWarn |
             Should -BeFalse
     }
 
     It 'should check granted inheritance flags' {
-        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ContainerAndLeaves | Should -BeTrue
-        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ChildLeaves  | Should -BeTrue
+        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ContainerAndLeaves -NoWarn |
+            Should -BeTrue
+        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ChildLeaves -NoWarn | Should -BeTrue
     }
 
 
     It 'should check exact ungranted inheritance flags' {
-        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ContainerAndLeaves -Exact |
+        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ContainerAndLeaves -Exact -NoWarn |
             Should -BeFalse
     }
 
     It 'should check exact granted inheritance flags' {
-        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ChildLeaves -Exact | Should -BeTrue
+        Test-CPermission @testDirPermArgs -Permission 'ReadAndExecute' -ApplyTo ChildLeaves -Exact -NoWarn | Should -BeTrue
     }
 
     It 'should check permission on private key' {
@@ -162,11 +169,11 @@ Describe 'Test-CPermission' {
                 $allPerm = 'GenericAll'
                 $readPerm = 'GenericRead'
             }
-            Grant-CPermission -Path $certPath -Identity $script:identity -Permission $allPerm
-            Test-CPermission -Path $certPath -Identity $script:identity -Permission $readPerm | Should -BeTrue
-            Test-CPermission -Path $certPath -Identity $script:identity -Permission $readPerm -Exact |
+            Grant-CPermission -Path $certPath -Identity $script:identity -Permission $allPerm -NoWarn
+            Test-CPermission -Path $certPath -Identity $script:identity -Permission $readPerm -NoWarn | Should -BeTrue
+            Test-CPermission -Path $certPath -Identity $script:identity -Permission $readPerm -Exact -NoWarn |
                 Should -BeFalse
-            Test-CPermission -Path $certPath -Identity $script:identity -Permission $allPerm, $readPerm -Exact |
+            Test-CPermission -Path $certPath -Identity $script:identity -Permission $allPerm, $readPerm -Exact -NoWarn |
                 Should -BeTrue
         }
         finally
@@ -186,8 +193,9 @@ Describe 'Test-CPermission' {
             Select-Object -First 1
         $cert | Should -Not -BeNullOrEmpty
         $certPath = Join-Path -Path 'cert:\' -ChildPath (Split-Path -NoQualifier -Path $cert.PSPath)
-        Get-CPermission -path $certPath -Identity $script:identity | Out-String | Write-Host
-        Test-CPermission -Path $certPath -Identity $script:identity -Permission 'FullControl' | Should -BeTrue
-        Test-CPermission -Path $certPath -Identity $script:identity -Permission 'FullControl' -Exact | Should -BeTrue
+        Get-CPermission -path $certPath -Identity $script:identity -NoWarn | Out-String | Write-Host
+        Test-CPermission -Path $certPath -Identity $script:identity -Permission 'FullControl' -NoWarn | Should -BeTrue
+        Test-CPermission -Path $certPath -Identity $script:identity -Permission 'FullControl' -Exact -NoWarn |
+            Should -BeTrue
     }
 }